Move setting of custom button dialog from ops_controller and infra_networking to a mixin
Code was copypasted from one to another and no change was needed so moved to mixin.
